refactor(billing-state): rename user-billing-state → org-billing-state
